rename: 캘린더 디렉토리 구조 변경
