+++ src/lib/CalendarContext.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import {createContext} from "react";
 
 const CalendarContext = createContext({});
